[Refactor] Move type info allocator adapter out of storage
