+++ be/src/exprs/agg/histogram.h
@@ -24,8 +24,8 @@
 #include "exprs/agg/aggregate.h"
 #include "exprs/agg/aggregate_traits.h"
 #include "gutil/casts.h"
+#include "runtime/type_info_allocator_adapter.h"
 #include "simdjson.h"
-#include "storage/type_info_allocator_adapter.h"
 #include "storage/types.h"
 
 namespace starrocks {
