+++ be/src/formats/parquet/statistics_helper.cpp
@@ -31,7 +31,6 @@
 #include "formats/parquet/schema.h"
 #include "gutil/casts.h"
 #include "storage/primitive/column_predicate_factory.h"
-#include "storage/types.h"
 #include "types/date_value.h"
 #include "types/logical_type.h"
 #include "types/type_descriptor.h"
