+++ be/src/runtime/type_info_allocator_adapter.h
