+++ be/src/storage/rowset/bitmap_index_writer.cpp
@@ -48,10 +48,10 @@
 #include "base/string/utf8.h"
 #include "fs/fs.h"
 #include "runtime/mem_pool.h"
+#include "runtime/type_info_allocator_adapter.h"
 #include "storage/primitive/rowid_types.h"
 #include "storage/rowset/encoding_info.h"
 #include "storage/rowset/indexed_column_writer.h"
-#include "storage/type_info_allocator_adapter.h"
 #include "storage/types.h"
 #include "types/olap_type_infra.h"
 #include "types/storage_type_traits.h"
