+++ be/src/storage/rowset/bloom_filter_index_writer.cpp
@@ -43,10 +43,10 @@
 #include "base/string/utf8.h"
 #include "common/bloom_filter.h" // for BloomFilterOptions, BloomFilter
 #include "fs/fs.h"
+#include "runtime/type_info_allocator_adapter.h"
 #include "storage/primitive/rowid_types.h"
 #include "storage/rowset/encoding_info.h"
 #include "storage/rowset/indexed_column_writer.h"
-#include "storage/type_info_allocator_adapter.h"
 #include "storage/types.h"
 #include "types/logical_type.h"
 #include "types/olap_type_infra.h"
