+++ be/src/storage/rowset/indexed_column_writer.cpp
@@ -43,6 +43,7 @@
 #include "common/config_rowset_fwd.h"
 #include "common/logging.h"
 #include "fs/fs.h"
+#include "runtime/type_info_allocator_adapter.h"
 #include "storage/primitive/key_coder.h"
 #include "storage/rowset/encoding_info.h"
 #include "storage/rowset/index_page.h"
@@ -50,7 +51,6 @@
 #include "storage/rowset/page_builder.h"
 #include "storage/rowset/page_io.h"
 #include "storage/rowset/page_pointer.h"
-#include "storage/type_info_allocator_adapter.h"
 #include "storage/types.h"
 
 namespace starrocks {
