+++ be/src/storage/schema_change_utils.cpp
@@ -23,8 +23,8 @@
 #include "runtime/exec_env.h"
 #include "runtime/mem_pool.h"
 #include "runtime/runtime_state.h"
+#include "runtime/type_info_allocator_adapter.h"
 #include "storage/chunk_helper.h"
-#include "storage/type_info_allocator_adapter.h"
 #include "types/bitmap_value.h"
 #include "types/hll.h"
 #include "types/percentile_value.h"
