+++ be/src/storage/tablet_reader.cpp
@@ -29,6 +29,7 @@
 #include "gen_cpp/tablet_schema.pb.h"
 #include "gutil/stl_util.h"
 #include "primary_key_encoder.h"
+#include "runtime/type_info_allocator_adapter.h"
 #include "storage/aggregate_iterator.h"
 #include "storage/base/merge_iterator.h"
 #include "storage/chunk_helper.h"
@@ -45,7 +46,6 @@
 #include "storage/seek_range.h"
 #include "storage/tablet.h"
 #include "storage/tablet_updates.h"
-#include "storage/type_info_allocator_adapter.h"
 #include "storage/types.h"
 #include "types/logical_type.h"
 #include "util/json_flattener.h"
