+++ be/test/storage/schema_change_test.cpp
@@ -28,6 +28,7 @@
 #include "gen_cpp/Exprs_types.h"
 #include "gtest/gtest.h"
 #include "runtime/descriptor_helper.h"
+#include "runtime/type_info_allocator_adapter.h"
 #include "storage/chunk_helper.h"
 #include "storage/convert_helper.h"
 #include "storage/delta_writer.h"
@@ -35,7 +36,6 @@
 #include "storage/storage_engine.h"
 #include "storage/tablet_manager.h"
 #include "storage/txn_manager.h"
-#include "storage/type_info_allocator_adapter.h"
 #include "testutil/column_test_helper.h"
 #include "testutil/schema_test_helper.h"
 #include "testutil/tablet_test_helper.h"
